backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
asyncpg
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
langchain
langchain-openai
langchain-community
langchain-ollama
chromadb
bcrypt==4.0.1
psycopg2-binary
openai
email-validator
langchain-text-splitters
